open-security-guardian/requirements.txt
# Open Security Guardian - Python Dependencies

# Django Framework
Django==4.2.13
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.2
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.7
redis==4.6.0
django-redis==5.3.0

# API & Serialization
drf-spectacular==0.26.4
celery==5.3.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Security & Authentication
django-oauth-toolkit==1.7.1
cryptography==41.0.3
python-jose[cryptography]==3.3.0

# HTTP Clients & Integrations
requests==2.31.0
httpx==0.24.1
pydantic==2.1.1
python-dotenv==1.0.0

# Data Processing
pandas==2.0.3
numpy==1.24.4
python-dateutil==2.8.2

# Vulnerability Scanner Integrations
python-nmap==0.7.1
tenable-io==1.17.0
qualysapi==0.6.0
python-libnmap==0.7.3

# Ticketing System Integrations
jira==3.5.0
python-servicenow==1.0.0

# Monitoring & Observability
prometheus-client==0.17.1
structlog==23.1.0
sentry-sdk==1.29.2

# Utilities
python-magic==0.4.27
validators==0.20.0
ipaddress==1.0.23
netaddr==0.8.0
PyYAML==6.0.1
Pillow==10.0.0

# Testing
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0
freezegun==1.2.2

# Development
black==23.7.0
isort==5.12.0
flake8==6.0.0
pre-commit==3.3.3
mypy==1.5.1
django-stubs==4.2.3

# Production
gunicorn==21.2.0
whitenoise==6.5.0
dj-database-url==2.0.0
